requirements.txt
google-genai
google-cloud-storage
Pillow
fastapi
uvicorn
python-multipart
supabase
google-cloud-pubsub